requirements.txt
argparse
validators
pwntools
xmltodict==0.12.0
yt-dlp<=2022.8.14